% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_weight.R
\name{aw_weight}
\alias{aw_weight}
\title{Output feature vector λ}
\usage{
aw_weight(data)
}
\arguments{
\item{data}{A data.frame after a two-by-two factor comparison.}
}
\value{
Feature vector λ.
}
\description{
Input a judgment matrix after a two-by-two factor comparison
to output feature vector λ.
}
\examples{
\donttest{
a <- c(1,3/4,1/2)
b <- c(4/3,1,1/3)
c <- c(2,3/2,1)
data <- data.frame(a,b,c)
aw_weight(data)
}
}
