% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_fmf.R
\name{gauss_fmf}
\alias{gauss_fmf}
\title{Gaussian fuzzy membership function}
\usage{
gauss_fmf(raster, sig, c, h)
}
\arguments{
\item{raster}{Raster* object.}

\item{sig}{Numeric.}

\item{c}{Numeric.}

\item{h}{Numeric.}
}
\value{
raster Raster* object..
}
\description{
Based on Gaussian fuzzy membership function to make the
value of the raster range between 0 and 1.
}
\details{
The formula of Gaussian fuzzy membership function is exp(-(x - c)^2/(2 * sig^2)).
}
\examples{
\donttest{
library(raster)
raster <- raster(ncols = 30, nrows = 15)
values(raster) <- runif(n = ncell(raster),min = 0, max = 100)
#Create a grid with values in the range of 0-100
plot(raster)
raster_gaussfmf <- gauss_fmf(raster = raster,sig = 50, c = 100, h=1)
plot(raster_gaussfmf[[1]])
}
}
