% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbell_fmf.R
\name{gbell_fmf}
\alias{gbell_fmf}
\title{Generalised bell fuzzy membership function}
\usage{
gbell_fmf(raster, a, b, c)
}
\arguments{
\item{raster}{Raster* object.}

\item{a}{Numeric.}

\item{b}{Numeric.}

\item{c}{Numeric.}
}
\value{
raster Raster* object..
}
\description{
Based on Generalised bell fuzzy membership function to make the
value of the raster range between 0 and 1.
}
\details{
The formula of generalised bell fuzzy membership function
function is 1/(1+(((x-c) a)^2)^b).
}
\examples{
\donttest{
library(raster)
raster <- raster(ncols = 30, nrows = 15)
values(raster) <- runif(n = ncell(raster),min = 0, max = 100)
#Create a grid with values in the range of 0-100
plot(raster)
raster_gbellfmf <- gbell_fmf(raster,a = 1, b = 2,c = 3)
#Make the value of the grid between 0 and 1
plot(raster_gbellfmf[[1]])
}
}
