% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fmf.R
\name{linear_fmf}
\alias{linear_fmf}
\title{Linear fuzzy membership function}
\usage{
linear_fmf(raster, a, b)
}
\arguments{
\item{raster}{Raster* object.}

\item{a}{Numeric.}

\item{b}{Numeric.}
}
\value{
raster Raster* object..
}
\description{
Based on Linear fuzzy membership function to make the
value of the raster range between 0 and 1.
}
\details{
The formula of linear fuzzy membership function is ax+b.
}
\examples{
\donttest{
library(raster)
raster <- raster(ncols = 30, nrows = 15)
values(raster) <- runif(n = ncell(raster),min = 0, max = 5)
#Create a grid with values in the range of 0-100
plot(raster)
raster_linearfmf <- linear_fmf(raster = raster,a = 0.2, b = 0)
plot(raster_linearfmf[[1]])
}
}
