% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min.R
\name{max_min}
\alias{max_min}
\title{Normalization of maximum and minimum values}
\usage{
max_min(raster)
}
\arguments{
\item{raster}{Raster* object.}
}
\value{
raster Raster* object(value is between 0 and 1).
}
\description{
The maximum-minimum normalization is the normalization process using the maximum and minimum values in the data column,
and the normalized value is between 0 and 1. It is calculated by making a difference between the
data and the minimum value of the column and dividing it by the extreme difference.
}
\details{
The specific formula is: x' = x - min/max - min
}
\examples{
\donttest{
library(raster)
raster <- raster(ncols = 30, nrow = 15)
values(raster) <- runif(ncell(raster),0,100)
#Normalization of maximum and minimum values
maxmin_raster <- max_min(raster)
plot(maxmin_raster)
}
}
