% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat.R
\name{oat}
\alias{oat}
\title{One-Factor-At-a-Time}
\usage{
oat(rasters, weights, range, step, output_dir)
}
\arguments{
\item{rasters}{Raster* object.}

\item{weights}{The weight of each factor should correspond to the order of the rasters.
The main factor needs to be ranked first.}

\item{range}{Range of factor weight adjustment.}

\item{step}{The step size of the factor weight adjustment.}

\item{output_dir}{Either a character string naming a file or a connection open for writing.
"" indicates output to the console.}
}
\value{
raster Raster* object.
}
\description{
By changing the weight value of only one factor at a time,
while keeping the other factors as constant as possible,
the degree and regularity of the effect of the change in the weight of
a single factor on the results is reflected.
}
\details{
The formula of One-Factor-At-a-Time is wi=(1-wa)*wi0/1-wa0 1≤i≤n,i≠a
wi0 is the initial weight of each risk factor and wa0 is the initial
weight of the major changing risk factor. The sum of all weights is equal to 1.
}
\examples{
\donttest{
library(raster)
#cattle as the main factor
rasters <- raster::stack("1_cattle.tif",
                         "2_pig.tif",
                         "3_sheep.tif")
#The order of the weights is aligned with the order of the rasters
weights <- c(0.2,0.3,0.5)
#Range of weight adjustment (\%)
range <- c(-10,10)
#Step size of weight adjustment (\%)
step <- 1
#All results are output in out_dir
oat(rasters = rasters,weights = weights,
range = range,step = step,output_dir = "oat//1_cattle")
}
}
