% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid_fmf.R
\name{sigmoid_fmf}
\alias{sigmoid_fmf}
\title{Sigmoid fuzzy membership function}
\usage{
sigmoid_fmf(raster, a, c)
}
\arguments{
\item{raster}{Raster* object.}

\item{a}{Number.}

\item{c}{Number.}
}
\value{
raster Raster* object..
}
\description{
Based on sigmoid fuzzy membership function to make the
value of the raster range between 0 and 1.
}
\details{
The formula of sigmoid fuzzy membership function is 1/(1+exp(-a*(raster-c))).
}
\examples{
\donttest{
library(raster)
raster <- raster(ncols = 30, nrows = 15)
values(raster) <- runif(n = ncell(raster),min = 0, max = 100)
plot(raster)
raster_sigmoidfmf <- sigmoid_fmf(raster = raster,a = 1, c = 2)
plot(raster_sigmoidfmf[[1]])
}
}
