% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertain.R
\name{uncertain}
\alias{uncertain}
\title{Uncertainty Analysis}
\usage{
uncertain(rasters)
}
\arguments{
\item{rasters}{Raster* object.}
}
\value{
raster Raster* object.
}
\description{
Standard deviation of all risk maps after adjusting the weights.
}
\details{
All the risk maps with adjusted weights are put together
and the to calculate the standard deviation (uncertainty map).
}
\examples{
\donttest{
library(terra)
#Read all risk maps after adjusting the weights.
rasters <- terra::rast(dir("//unceratin_map",full.names = TRUE))
unceratin_map <- uncertain(rasters)
terra::plot(unceratin_map)
}
}
