% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_fun.R
\name{weight_fun}
\alias{weight_fun}
\title{Output the weights of the factors}
\usage{
weight_fun(data)
}
\arguments{
\item{data}{A data.frame after a two-by-two factor comparison.}
}
\value{
A decision matrix and weights of all factors (the sum of all weights is 1).
}
\description{
Input a judgment matrix after a two-by-two factor comparison
to output the weights of all factors (the sum of all weights is 1).
}
\examples{
\donttest{
a <- c(1,3/4,1/2)
b <- c(4/3,1,1/3)
c <- c(2,3/2,1)
data <- data.frame(a,b,c)
weight_fun(data)
}
}
