% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlc.R
\name{wlc}
\alias{wlc}
\title{Weighted linear combination}
\usage{
wlc(rasters, weights)
}
\arguments{
\item{rasters}{Raster* object.}

\item{weights}{Numeric. The weight of each factor should correspond to the order of the rasters.}
}
\value{
raster Raster* object.
}
\description{
Input the raster and weight values for each factor to
output a decision map.
}
\details{
The formula of weighted linear combination function is ax1+bx2+cx3...,
the weights a, b, c.... of factors x1, x2 and x3... of the sum of 1
}
\examples{
\donttest{
library(raster)

# Create a stack of rasters
rasters <- stack("1_cattle.tif", "2_pig.tif", "3_sheep.tif")

# Define the weights
#The order of the weights is aligned with the order of the rasters
weights <- c(0.2,0.3,0.5)

# Ensure that the variable names are consistent
result <- wlc(rasters, weights)

# Plot result
plot(result)
}
}
