% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_weight.R
\name{aw_weight}
\alias{aw_weight}
\title{Calculate Feature Vector Lambda from a Judgment Matrix}
\usage{
aw_weight(data)
}
\arguments{
\item{data}{A \code{data.frame} after a two-by-two factor comparison.}
}
\value{
A list containing the feature vector lambda, the product of the matrix and weight factors,
and the normalized values of these products.

A list containing the feature vector lambda, the product of the matrix and weight factors,
and the normalized values of these products.
}
\description{
Calculate Feature Vector Lambda from a Judgment Matrix
}
\details{
This function processes a judgment matrix obtained from pairwise factor comparisons to compute the feature vector lambda, which is an indicator of the consistency of the matrix. The function returns the feature vector lambda, detailed calculations of the matrix product with weights, and the division of these products by their corresponding weights.
}
\examples{
\donttest{
a <- c(1, 3/4, 1/2)
b <- c(4/3, 1, 1/3)
c <- c(2, 3/2, 1)
data <- data.frame(a, b, c)
lambda_result <- aw_weight(data)
print(lambda_result)
}
Calculate Feature Vector Lambda from a Judgment Matrix

This function processes a judgment matrix obtained from pairwise factor comparisons
to compute the feature vector lambda, which is an indicator of the consistency of the matrix.
The function returns the feature vector lambda, detailed calculations of the matrix product 
with weights, and the division of these products by their corresponding weights.

\donttest{
a <- c(1, 3/4, 1/2)
b <- c(4/3, 1, 1/3)
c <- c(2, 3/2, 1)
data <- data.frame(a, b, c)
lambda_result <- aw_weight(data)
print(lambda_result)
}
}
