% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consist_test.R
\name{consist_test}
\alias{consist_test}
\title{Perform Consistency Test on Judgment Matrix}
\usage{
consist_test(lambda, n)
}
\arguments{
\item{lambda}{Numeric, the feature vector lambda calculated by \code{aw_weight()}.}

\item{n}{Integer, the number of factors used in the matrix.}
}
\value{
Numeric, the Consistency Ratio (CR).
}
\description{
This function takes the lambda value obtained from a judgment matrix and the number of factors to evaluate the consistency of the matrix. The consistency ratio (CR) is calculated and used to determine if the matrix meets the threshold of consistency.
}
\details{
The consistency test is essential in the Analytic Hierarchy Process (AHP) to ensure
that the judgment matrix is reasonably consistent. A CR less than or equal to 0.10 indicates
acceptable consistency.
The reference values for the Random Index (RI) used are: c(0,0,0.52,0.89,1.12,1.26,1.36,1.41,1.46,1.49,1.52,1.54,1.56,1.58,1.59).
}
\examples{
\donttest{
lambda <- 1.5
n <- 10
CR_result <- consist_test(lambda, n)
print(CR_result)
}
}
