% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_fmf.R
\name{gauss_fmf}
\alias{gauss_fmf}
\title{Normalize Raster Values Using Gaussian Fuzzy Membership Function}
\usage{
gauss_fmf(raster, sig, c, h)
}
\arguments{
\item{raster}{Raster* object representing the raster to be transformed.}

\item{sig}{Numeric value for the standard deviation of the Gaussian function.}

\item{c}{Numeric value representing the center of the Gaussian curve.}

\item{h}{Numeric height scaling factor.}
}
\value{
Raster* object with transformed values.
}
\description{
This function applies a Gaussian fuzzy membership function to a raster object to scale its values between 0 and 1 based on the specified parameters of the Gaussian function.
}
\details{
The formula of Gaussian fuzzy membership function is exp(-(x - c)^2/(2 * sig^2)).
}
\examples{
\donttest{
library(raster)
# Create a raster with random values between 0 and 100
raster_obj <- raster(ncols = 30, nrows = 15)
values(raster_obj) <- runif(ncell(raster_obj), min = 0, max = 100)
plot(raster_obj)
# Apply Gaussian fuzzy membership function
transformed_raster <- gauss_fmf(raster = raster_obj, sig = 50, c = 50, h = 1)
plot(transformed_raster)
}
}
