% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbell_fmf.R
\name{gbell_fmf}
\alias{gbell_fmf}
\title{Normalize Raster Values Using Generalized Bell Fuzzy Membership Function}
\usage{
gbell_fmf(raster, a, b, c)
}
\arguments{
\item{raster}{Raster* object representing the raster to be transformed.}

\item{a}{Numeric, width parameter of the bell curve.}

\item{b}{Numeric, slope parameter of the bell curve.}

\item{c}{Numeric, center of the bell curve.}
}
\value{
Raster* object with values scaled between 0 and 1.
}
\description{
This function scales raster values to the range \link{0, 1} using the Generalized Bell Fuzzy Membership Function. This transformation is useful for fuzzy logic operations and other applications where data scaling is necessary.
}
\details{
The formula of generalised bell fuzzy membership function is 1/(1+(((x-c) a)^2)^b).
}
\examples{
\donttest{
library(raster)
# Create a raster with random values
raster_obj <- raster(ncols = 30, nrows = 15)
values(raster_obj) <- runif(ncell(raster_obj), min = 0, max = 100)
# Apply the Generalized Bell Fuzzy Membership function
transformed_raster <- gbell_fmf(raster_obj, a = 2, b = 3, c = 50)
# Visualize the transformed raster
plot(transformed_raster)
}
}
