% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macr.R
\name{macr}
\alias{macr}
\title{Mean of Absolute Change Rates}
\usage{
macr(oat_rasters, wlc_raster, output_dir, factor_name)
}
\arguments{
\item{oat_rasters}{A RasterStack object containing rasters from OFAT analysis for various factors.}

\item{wlc_raster}{A RasterLayer object that represents the baseline or weighted linear combination from which deviations are calculated.}

\item{output_dir}{The directory path where the output files will be saved. This path must be accessible and writable.}

\item{factor_name}{A string specifying the name of the main factor. This name is used to organize the output files into a dedicated subdirectory within the output directory, ensuring that the results are easy to locate and review.}
}
\value{
Does not return a value but writes a CSV file with detailed statistical analysis results, along with individual change rate raster files.
}
\description{
Computes enhanced statistical metrics for the results from a One-Factor-at-a-Time (OAT) sensitivity analysis of a main factor. This function aims to provide a deeper insight into how changes in a factor's weight affect the model's outcomes by analyzing mean, standard deviation, minimum, and maximum absolute change rates.
}
\details{
This function enhances the traditional Mean of Absolute Change Rates (MACR) analysis by not only calculating the mean change rate but also providing additional statistics such as the standard deviation, minimum, and maximum change rates. These metrics offer a more comprehensive view of the variability and distribution of change rates, helping to identify how stable or volatile the results are under different weight adjustments. The function requires a RasterStack of OFAT results, a baseline raster to compare against, and outputs the results in a well-organized CSV file within a directory named after the analyzed factor.
}
\examples{
\donttest{
# Calculate the mean of absolute change rates for the factor "1_AH".
library(raster)
# Read the results of the One-Factor-at-a-Time analysis for the factor "AH".
oat_rasters <- raster::stack(dir("//oat//1_AH", full.names = TRUE))
# Read the decision map obtained with the weighted linear combination.
wlc_raster <- raster::raster("//wlc//decision_map.tif")
# Specify the output directory.
output_dir <- "//macr"
# Calculate enhanced MACR and save the results under the specified factor's directory.
macr(oat_rasters = oat_rasters, wlc_raster = wlc_raster,
     output_dir = output_dir, factor_name = "1_AH")
}
}
