% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oat.R
\name{oat}
\alias{oat}
\title{One-Factor-At-a-Time Sensitivity Analysis}
\usage{
oat(rasters, weights, range, step, output_dir, plot = FALSE, dominant_factor)
}
\arguments{
\item{rasters}{RasterStack object containing the layers to be analyzed.}

\item{weights}{Numeric vector, initial weights of each raster layer, must correspond to the order in the raster stack.}

\item{range}{Numeric vector of length 2, defining the range (min, max) of weight adjustment in percentage points.}

\item{step}{Numeric, the increment in weight adjustment across the specified range.}

\item{output_dir}{Character string, path to the directory where output rasters will be saved.}

\item{plot}{Logical, whether to plot the raster differences using \code{rasterVis::levelplot}.}

\item{dominant_factor}{Character string, the name of the dominant factor.}
}
\value{
Returns nothing but writes adjusted raster files and optionally plots them.
}
\description{
Performs sensitivity analysis by adjusting the weight of one factor at a time
and assessing the impact on the resulting raster. This helps in understanding the influence
of each factor on the outcome while other factors are held constant.
}
\details{
Implements the One-Factor-At-a-Time method where each factor's weight is adjusted
individually and the effect is observed in the output. The formula used is wi=(1-wa)*wi0/(1-wa0) for 1≤i≤n, i≠a.
wi0 is the initial weight of each risk factor, wa0 is the initial weight of the adjusted risk factor.
The sum of all weights equals 1. This function also calculates the difference from the baseline scenario
where no weights are adjusted and can plot the results if desired.
}
\examples{
\donttest{
library(raster)
library(rasterVis)

# Assume there are three raster files representing different risk factors
# For example, these raster files might be located in a subdirectory of your working directory
rasters <- raster::stack(
  system.file("extdata", "1_AH.tif", package = "SpatMCDA"),
  system.file("extdata", "2_Climate.tif", package = "SpatMCDA"),
  system.file("extdata", "3_CV.tif", package = "SpatMCDA")
)

# Assign names to each raster layer, which is necessary for specifying the dominant_factor
names(rasters) <- c("1_AH", "2_Climate", "3_CV")

# Set initial weights, ensuring the order of weights matches the order of raster layers
weights <- c("1_AH" = 0.2, "2_Climate" = 0.3, "3_CV" = 0.5)

# Set the range and step size for weight adjustments
range <- c(-20, 20)  # Percentage range for weight adjustment (\%)
step <- 1            # Step size for weight adjustment (\%)

# Set the output directory, ensure this directory exists or you have permission to create files in it
output_dir <- "/oat"

# Call the improved oat() function
# Specify 'pig' as the main factor to adjust
oat(
  rasters = rasters,
  weights = weights,
  dominant_factor = "1_AH"    # Specify '1_AH' as the dominant factor
  range = range,
  step = step,
  output_dir = output_dir,
  plot = TRUE               # Set to TRUE if you want to see visualizations of the differences
)
}
}
