% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_macr.R
\name{plot_macr}
\alias{plot_macr}
\title{Plot the Mean of Absolute Change Rates}
\usage{
plot_macr(data)
}
\arguments{
\item{data}{data.frame, a data frame that must include the following columns:
\itemize{
\item CRW: a numeric vector representing the change rate of weights (\%) for each factor.
\item Value: a numeric vector representing the mean of absolute change rates (\%) obtained from enhanced_macr() function or similar analyses.
\item Variable: a factor vector representing the names of the factors. Each factor name correlates with CRW and Value.
}}
}
\value{
A ggplot object that plots the mean of absolute change rates against the change rate of weights for each factor, differentiated by colors and shapes for each factor.
}
\description{
Plots the mean of absolute change rates for various factors across different weight adjustments. This function visualizes the sensitivity of model outcomes to changes in factor weights, providing a clear visual representation of how different weight adjustments impact the calculated change rates.
}
\examples{
\donttest{
# Prepare data for three factors with their corresponding change rates and MACRs
CRW <- c(-5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5)
Value_A <- c(0.2273, 0.1819, 0.1364, 0.0909, 0.0455, 0, 0.0455, 0.0909, 0.1364, 0.1819, 0.2273)
Value_B <- c(0.2213, 0.177, 0.1328, 0.0885, 0.0443, 0, 0.0443, 0.0885, 0.1328, 0.177, 0.2213)
Value_C <- c(0.11065, 0.08850, 0.06640, 0.04425, 0.02215, 0, 0.02215, 0.04425, 0.06640, 0.08850, 0.11065)
Factor_A <- rep("Factor_A", 11)
Factor_B <- rep("Factor_B", 11)
Factor_C <- rep("Factor_C", 11)

# Combine into a data frame
macr_data <- data.frame(CRW = rep(CRW, 3), Value = c(Value_A, Value_B, Value_C),
                        Variable = factor(c(Factor_A, Factor_B, Factor_C)))

# Plot the Mean of Absolute Change Rates
plot_macr(macr_data)
}
}
