% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight.R
\name{plot_weight}
\alias{plot_weight}
\title{Visualize Weights of Factors}
\usage{
plot_weight(data, title = "Weights of Factors")
}
\arguments{
\item{data}{Numeric vector containing the weights of all factors obtained from \code{weight_fun()}.}

\item{title}{Optional; the title of the plot.}
}
\value{
Invisible returns the barplot object, primarily used for its side effect of plotting.
}
\description{
Creates a horizontal bar plot visualizing the weights of factors as calculated by the \code{weight_fun()} function. This plot helps in understanding the relative importance of each factor in a decision-making model.
}
\examples{
\donttest{
a <- c(1, 3/4, 1/2)
b <- c(4/3, 1, 1/3)
c <- c(2, 3/2, 1)
data_frame <- data.frame(a, b, c)
weight_factors <- weight_fun(data_frame)$weight_factor
plot_weight(weight_factors, "Factor Weight Distribution")
}
}
