% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlc.R
\name{wlc}
\alias{wlc}
\title{Perform Weighted Linear Combination on Raster Layers}
\usage{
wlc(rasters, weights)
}
\arguments{
\item{rasters}{A \code{terra::SpatRaster} object or a compatible object from the \code{raster} package.}

\item{weights}{Numeric vector specifying the weights for each raster layer.
The weights should correspond in order to the layers in the \code{rasters} object.}
}
\value{
A \code{terra::SpatRaster} object representing the weighted combination of input rasters.
}
\description{
This function applies a weighted linear combination to a stack of raster layers to produce a decision map. Each raster is multiplied by a corresponding weight and the results are summed to create a single output raster. This method is commonly used in spatial decision-making processes.
}
\details{
The weighted linear combination (WLC) formula is:
\deqn{a \cdot x_1 + b \cdot x_2 + c \cdot x_3 + \ldots}
where \code{a}, \code{b}, \code{c},... are the weights for factors \code{x_1}, \code{x_2}, \code{x_3},... respectively.
The sum of all weights should be 1.
}
\examples{
\donttest{
library(terra)

# Assuming rasters are loaded or created here
rasters <- rast(c(system.file("extdata", "1_AH.tif", package = "SpatMCDA"),
                  system.file("extdata", "2_Climate.tif", package = "SpatMCDA"),
                  system.file("extdata", "3_CV.tif", package = "SpatMCDA")))

# Define the weights
weights <- c(0.2, 0.3, 0.5)

# Perform the weighted linear combination
result <- wlc(rasters, weights)

# Plot result
plot(result)
}
}
